Tidy decorator, move to toml module
